+++ src/App.js
@@ -2,9 +2,11 @@ import React from 'react';
 import Sidebar from './components/Sidebar';
 import './App.css';
 import Home from './components/pages/home/Home';
-import About from './components/pages/About';
-import GetInvolved from './components/pages/GetInvolved';
+import About from './components/pages/about/About';
+import GetInvolved from './components/pages/about/GetInvolved';
 import Partners from './components/pages/Partners';
+import Event from './components/pages/Event';
+import Contact from './components/pages/Contact';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 function App() {
@@ -18,6 +20,8 @@ function App() {
               <Route path='/about' element={<About />} />
               <Route path='/get-involved' element={<GetInvolved />} />
               <Route path='/partners' element={<Partners />} />
+              <Route path='/events' element={<Event />} />
+              <Route path='/contact' element={<Contact />} />
           </Routes>
         </div>
       </Router>
