--- src/components/pages/History.js
@@ -1,32 +1,0 @@
-import React from "react";
-
-const History = () => {
-  return (
-    <div className="pt-5">
-      <h5>Tracing Our Digital Footprints</h5>
-      <p>Our origin is rooted in the passion of young Afghans eager to make a digital difference.</p>
-      <ul>
-        <li>
-          In 2017, a spark was ignited during the "Youth in Internet Governance" workshop at
-          the first National IGF Afghanistan (IGFA). That spark expanded, involving
-          stakeholders from diverse arenas.
-        </li>
-        <br />
-        <li>
-          By 2018, our commitment to the cause bore fruit with the establishment of
-          pre-YouthIGFA, exclusively centering the discourse around youth and Internet
-          Governance.
-        </li>
-        <br />
-        <li>
-          Finally, in 2019 we were able to organize the first full-fledged Youth Internet
-          Governance Forum Afghanistan (YouthIGFA).
-        </li>
-      </ul>
-      <p>Every milestone is a testament to the spirit of Afghan youth, their dedication, and their
-      undying belief in a better digital future.</p>
-    </div>
-  );
-}
-
-export default History;
