--- src/components/pages/teamData.js
@@ -1,143 +1,0 @@
-import User from "../../assets/user.jpg";
-
-const team = [
-  {
-    organizingCommittee: [
-      {
-        name: "Ghayoor Bawary",
-        position: "Chairperson",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Mohib Utmankhil",
-        position: "Vice Chairperson",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Qudratullah Hiwadpal",
-        position: "Program Manager",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Lima Madomi",
-        position: "Marketing & Outreach Lead",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Geety Mahdizada",
-        position: "Logistics Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Sajia Yarmal",
-        position: "Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-    ],
-    secretariat: [
-      {
-        name: "Sajia Yarmal",
-        position: "Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Rooh Afza Abubakr",
-        position: "Program Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Khpal Wak Shinwari",
-        position: "Operations Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Masuma Rasuli",
-        position: "Social Media/Communication Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Ahmad Zaki Yousufi",
-        position: "Web Coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Somebody",
-        position: "Reports coordinator",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-    ],
-    advisoryCommittee: [
-      {
-        name: "Omar Ansari",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Amrita Choudhury",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Jenna Manhau Fung",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Jamal Noorzaie",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Syed Hashimi",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-    ],
-    speakersAndTrainers: [
-      {
-        name: "Abdul Azim Raheemi",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Mohib Utmankhil",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Lima Madomi",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Zainab Bahar Amini",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Mirwais Khan",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-      {
-        name: "Atiqullah Ahmadzai",
-        image: User,
-        bio: "The shimmering stars danced in the velvety night sky, casting a magical glow over the serene landscape. Whispers of the wind carried secrets of forgotten tales, as the moonlight painted ethereal patterns on the tranquil meadow. Nature's symphony whispered melodies of enchantment, captivating hearts with its timeless beauty"
-      },
-    ],
-  },
-];
-
-export default team;
